Update Junk.java
cambiato attributo a double

+++ src/main/java/it/project1/items/Junk.java
@@ -2,7 +2,7 @@ package it.project1.items;
 
 public class Junk extends Item implements IUsable, ISellable{
 
-    public Junk(String name, String description, int value, int weight, String rarity, int requirments, boolean isStackable, int durability) {
+    public Junk(String name, String description, int value, double weight, String rarity, int requirments, boolean isStackable, int durability) {
         super(name, description, value, weight, rarity, requirments, isStackable, durability);
     }
